modify bug

+++ lib/lib_net/server.cpp
@@ -1,7 +1,7 @@
 #include "net.h"
 
 int main(int argc, char ** argv){
-    netServer *netserver = new netServer(50000);
+    netServer *netserver = new netServer((unsigned short)50000);
     netserver->initialize();
     netserver->start();
     return 0;
